refactor(demo-web): relocate `useBrowserLanguage` to shared `lib/hooks` directory
Move `useBrowserLanguage` hook from `features/upload` to a shared `lib/hooks` directory for improved reusability across components. Update all relevant imports to reflect the new path.
